Move Apps to its own file
